requirements.txt
python-dotenv
rich
addict
pytelegrambotapi
groq
web3